output: Extract file review footer
File-review output now has separate modules for row rendering and shared
summary labels. The main renderer still carries footer command hint
assembly and temporary review-state construction.

That footer code imports command builders and selected-change state
records. Keeping those dependencies in the page renderer makes the output
package look like the page loop owns command construction rather than only
delegating to it.

This commit moves footer command rendering into output.file_review_footer
and delegates the footer call from output.file_review.

Subsequent commits will add a direct architecture guard for the footer
module.
